bcache-tools: Export CACHED_UUID and CACHED_LABEL
From: Ryan Harper <ryan.harper@canonical.com>
https://github.com/koverstreet/bcache-tools/pull/1

Add bcache-export-cached helper to export CACHED_UUID and CACHED_LABEL always

Linux kernel bcache driver does not always emit a uevent[1] for when
a backing device is bound to a bcacheN device.  When this happens, the udev
rule for creating /dev/bcache/by-uuid or /dev/bcache/by-label symlinks does
not fire and removes any persistent symlink to a specific backing device
since the bcache minor numbers (bcache0, 1, 2) are not guaranteed across reboots.

This script reads the superblock of the bcache device slaves,ensuring the slave
is a backing device via sb.version check, extracts the dev.uuid and
dev.label values and exports them to udev for triggering the symlink rules in
the existing rules file.

1. https://bugs.launchpad.net/ubuntu/+source/linux/+bug/1729145

Signed-off-by: Shaoxiong Li <dahefanteng@gmail.com>